Add a scheduled workflow to synchronize branches from `gitster/git`
The Git maintainer publishes only the integration branches to `git/git`,
leaving the individual contribution branches in `gitster/git`.

We do want to make sure that the latter branches are available, and
synchronized, in `gitgitgadget/git`, to facilitate contributions to
patch series that are in flight.

So let's synchronize the branches, making sure that the integration
branches (`maint`, `master`, `next`, `seen`) are not synchronized from
`gitster/git`; They are synchronized via the `sync-ref` workflow
already.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>