sync-gitster-git-branches: limit concurrency
It does not make sense to try synchronizing in parallel.

I considered adding `cancel-in-progress: true`, but that strikes me as
wasteful: if we're already pushing an update, canceling it would require
the entire work to be done again, while chances are that any queued
run will likely find that everything is already up to date and no
further work needs to be done.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>