Use the official GitHub Action to obtain installation access tokens
When I committed the first workflow in e84a016 (Add a workflow to
synchronize refs from git/git to gitgitgadget/git, 2023-08-18), I had
missed the first public release of `actions/create-github-app-token` by
_mere days_: The first version of that GitHub Action to be officially
published to the GitHub Marketplace was v1.1.0, on August 22nd, 2023:
https://github.com/actions/create-github-app-token/releases/tag/v1.1.0,
i.e. a mere 4 days after I had committed the initial version of the
`sync-ref` workflow.

That was the reason why, when I scoured around for a GitHub Action that
could be used to obtain an installation access token, I only found a
third-party one.

I don't exactly remember how I got aware of the official Action provided
by GitHub itself, but I know that I immediately opened
https://github.com/gitgitgadget/gitgitgadget-workflows/issues/5 so that
I wouldn't forget to eventually migrate to that Action.

This commit is it. With this, we switch from that 3rd-party Action to
the official GitHub one.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>