handle-pr-comment/handle-pr-push: create a Check Run in the PR
One of the really nice things about the Azure Pipelines was that they
would automatically show PR Checks when they were running. This gave
some sort of pleasing, visual feedback when `/submit`-ing patch series.

Let's have the same even in the new world, where the PR comment is
handled by a GitHub workflow running in a completely separate GitHub
org, and not connected with the PR in any official capacity.

Note: If we do not specify `output[text]` in the final update, the text
that was specified when creating the Check Run will be _forgotten_. This
is arguably a bug in Actions, and given the recent history of that
product it probably won't get fixed, so we work around it here by simply
specifying the entire `output` _again_.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>