Prevent a the new workflows from running concurrently
The newly-added workflows update global state. This global state is held
in Git notes in the `gitgitgadget/git` repository and is updated in
reaction to updates to the Git mailing list mirror and to the `seen`
branch.

While there is now logic in GitGitGadget's Git notes handling to allow
for "distributed updates", it makes more sense to hold off from trying
to perform (nearly-)identical updates in parallel.

Note: In the case of `handle-pr-comment` and `handle-pr-push`, let's be
mindful that it should be possible to handle different PRs concurrently,
which is definitely a change of behavior from the current Azure
Pipelines setup, where only one job can be run at any given time,
_across_ Azure Pipelines (for example, a lengthy `Update GitGitGadget's
PRs` run can easily delay any `/submit` from being processed).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>