keepalive: in forks, merge the upstream branch instead of creating commits
In the fork in `gitgitgadget-workflows/gitgitgadget-workflows`, we now
have a divergent commit history. That is because both that fork as well
as the upstream repository in `gitgitgadget/gitgitgadget-workflows`
create those keepalive commits.

That is not ideal.

Let's instead merge the upstream branch in forked repositories, which
will include that already-created keepalive commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>