keepalive: run twice a month
The reason for the keepalive workflow is that scheduled workflows are
disabled in repositories that are "not active". The
`gitgitgadget-workflows` repository, however, _is_ active, even if there
are no code changes. In fact, the hope is that there won't be any need
for code changes at some stage because the code is perfect, and only
runs the workflows as orchestrated by GitGitGadget's GitHub App as well
as the scheduled workflows.

Now that we changed the strategy in forked repositories from creating
keepalive commits to first trying to grab a keepalive commit from the
upstream repository, we need to increase the frequency of trying to do
so because the `gitgitgadget` org's runners are unfortunately frequently
very busy (the Git maintainer delights in pushing out dozens of topic
branches simultaneously, each of which triggers CI builds with a
combined runtime of over 8 hours...), which means that the `keepalive`
workflow might run in the fork well before the upstream repository gets
to creating the keepalive commit.

So let's just run things twice a month instead of just once a month.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>