sync-git-gui: rename it, to form the basis of a generic workflow
In the upcoming work to support a project other than Git, we will need
to synchronize a different upstream repository to a different "PR
repository".

To that end, let's consolidate `sync-git-gui` and `sync-gitster-git`
into a single, matrix workflow that will continue to synchronize the Git
GUI and the fine-grained branches from `gitster/git` to
`gitgitgadget/git`, but can be overridden in another fork of
`gitgitgadget-workflows` via a repository variable that contains the
project config to perform the equivalent job of synchronizing upstream
branches into the repository where GitGitGadget handles the Pull
Requests for said project.

This is step 1 to support that: renaming the `sync-git-gui` workflow to
`sync-upstream-branches`. Subsequent commits will transmogrify that
workflow into the desired shape.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>