sync-upstream-branches: take over the job of the `sync-gitster-git` workflow
With a few touch-ups, the workflow can become generic enough to handle
also the synchronization of `gitster/git`'s branches to
`gitgitgadget/git`.

The overall goal, of course, is to prepare for supporting projects other
than Git, via configurations stored in repository variables. That's the
task of the next commit(s).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>